+++ src/builtins/abbr.rs
@@ -38,6 +38,7 @@ struct Options {
     set_cursor_marker: Option<WString>,
     args: Vec<WString>,
     color: ColorEnabled,
+    description: Option<WString>,
 }
 
 impl Options {
@@ -146,6 +147,10 @@ fn abbr_show(opts: Options, streams: &mut IoStreams, parser: &mut Parser) -> Bui
             if let Some(ref set_cursor_marker) = abbr.set_cursor_marker {
                 add_arg(L!("--set-cursor="));
                 add_arg(&escape_string(set_cursor_marker, style));
+            }
+            if let Some(ref description) = abbr.description {
+                add_arg(L!("--description="));
+                add_arg(&escape_string(description, style));
             }
             if abbr.replacement_is_function {
                 add_arg(L!("--function"));
@@ -418,6 +423,7 @@ fn abbr_add(opts: Options, streams: &mut IoStreams) -> BuiltinResult {
             set_cursor_marker: opts.set_cursor_marker,
             from_universal: false,
             commands: opts.commands,
+            description: opts.description,
         });
     });
 
@@ -481,6 +487,7 @@ pub fn abbr(parser: &mut Parser, streams: &mut IoStreams, argv: &mut [&wstr]) ->
             ArgType::OptionalArgument,
             SET_CURSOR_SHORT,
         ),
+        wopt(L!("description"), ArgType::OptionalArgument, 'd'),
         wopt(L!("function"), ArgType::RequiredArgument, 'f'),
         wopt(L!("rename"), ArgType::NoArgument, RENAME_SHORT),
         wopt(L!("erase"), ArgType::NoArgument, 'e'),
@@ -549,6 +556,9 @@ pub fn abbr(parser: &mut Parser, streams: &mut IoStreams, argv: &mut [&wstr]) ->
                     return Err(STATUS_INVALID_ARGS);
                 }
                 opts.regex_pattern = w.woptarg.map(ToOwned::to_owned);
+            }
+            'd' => {
+                opts.description = w.woptarg.map(ToOwned::to_owned);
             }
             SET_CURSOR_SHORT => {
                 if opts.set_cursor_marker.is_some() {